Float型に対応しました。
refactoring:  triangle_Spec
内容: もう少しSpecに同値分割と境界値分析を使用

+++ triangle.rb
@@ -1,7 +1,7 @@
 class Triangle
 
   def self.pattern(a, b, c)
-    return '正しい値を入力してください。' if !a.kind_of?(Integer) or !b.kind_of?(Integer) or !c.kind_of?(Integer)
+    return '正しい値を入力してください。' if !a.kind_of?(Numeric) or !b.kind_of?(Numeric) or !c.kind_of?(Numeric)
 
     return '三角形じゃないです＞＜' unless a < (b + c) and b < (a + c) and c < (a + b)
     return '正三角形ですね！' if a == b and b == c
@@ -10,5 +10,5 @@ class Triangle
   end
 end
 
-puts Triangle.pattern(ARGV[0].to_i, ARGV[1].to_i, ARGV[2].to_i)
+puts Triangle.pattern(ARGV[0].to_f, ARGV[1].to_f, ARGV[2].to_f)
